Switching "int" to "long int" where needed in the diffusion handler. Removing unnecessary physics from the solver for the UO2 case. Removing the compiling of the tests. SB 20160908
git-svn-id: file:///Users/MacAir/Dropbox/xolotl-cleaned/branches/xenon-64-branch@1053 72105f29-0fe1-43bd-9363-b1e281e962a2

+++ xolotl/xolotlCore/diffusion/Diffusion1DHandler.cpp
@@ -94,7 +94,7 @@ void Diffusion1DHandler::computeDiffusion(IReactionNetwork *network,
 
 void Diffusion1DHandler::computePartialsForDiffusion(
 		IReactionNetwork *network,
-		double *val, int *indices, double hxLeft, double hxRight, int ix,
+		double *val, long int *indices, double hxLeft, double hxRight, int ix,
 		double, int, double, int) {
 	// Get all the reactants
 	auto reactants = network->getAll();
